#191: extract PendingFilesSummary to own file
Per repo convention — prefer one top-level class per file.
