+++ composeApp/src/commonMain/kotlin/com/tubetoast/tether/presentation/PendingFilesSummary.kt
@@ -1,0 +1,10 @@
+package com.tubetoast.tether.presentation
+
+data class PendingFilesSummary(
+    val fileCount: Int,
+    val totalBytes: Long,
+) {
+    companion object {
+        val NONE = PendingFilesSummary(0, 0L)
+    }
+}
